Add :input option to FFI::TypesGenerator#generate
This is useful when while porting to new achitectures, you are not able to
run gcc directly from a ssh connection (e.g. gcc is only available on a
chroot), so you run gcc manually against the input that TypesGenerator
would give to gcc, copy back the resulting C file to your local
machine, and pass the C file to be parsed by TypesGenerator.