Fix variadic calls with float/double arguments
Problem was due to incorrect way to call C method ffi_prep_cif_var,
where third argument was passed as total number of arguments instead of
fixed number of arguments

This fixes ruby-ffi on the ppc64el architecture, and should be harmless
on other architectures. This was tested at least on amd64 (x86_64) and
ppc64el.

Signed-off-by: Antonio Terceiro <terceiro@debian.org>