Add changes for nested structure support
The "initialize" method for FFI::Struct is written in C, so writing a
Ruby version causes problems.  NiceFFI avoided this by using a sub-
class and calling 'super'.  I don't want to require a sub-class here
and I don't want to risk breaking the C code, so eliminate the
constructor-based data import and make import a separate operation
instead.