remove RB_GC_GUARD_PTR references
RB_GC_GUARD_PTR is an 'internal' API and consumers should only use
RB_GC_GUARD.  On non-GCC/non-MSC compilers (i.e. "Sun" Compiler on
Solaris or xlc compiler on AIX) this macro does not exist in a usable
form.

In this case both of the instances of RB_GC_GUARD_PTR were unnecessary
since argv was already protected and rbParams was ALLOCA_N'd onto the
stack which does not require protection.