Add tests for calls between Fiddle and FFI
This is particular interesting because FFI stores thread local call info
on the stack, which is retrieved when a callback is received.
This is something that Fiddle doesn't do, but a callback should be handled
gracefully nevertheless.

The test "from fiddle to ffi" fails without the previous commit.

These tests succeed on MRI and on JRuby (provided
https://github.com/jruby/jruby/pull/4518 will be merged).