Query ruby thread and GVL states instead of relying on our call frame for callbacks
This fixes #527 and other interoperability issues that can happen,
when a callback is called without a valid call frame.

The change is enabled for ruby-2.3+ only, since it requires the
functions ruby_native_thread_p() and ruby_thread_has_gvl_p().

The creation of a call frame is not (yet) removed, because
it still carries callback expections around.