read_string should not throw an error on length 0
Often strings with length 0 are accompanied by a pointer to an invalid
memory address.

struct alert {
   uint error_code;
   uint message_length;
   char *message;
};
```

If the message length is 0 then message pointer is often also 0 - i.e.
a null pointer.

Before this a check would be required for length before calling
read_string as otherwise it will raise an exception instead of
returning an empty string as expected.