Revert "Do closures via libffi"
This reverts commit 9f257af19e6088c2986f85dea17455e52a2be405.

This is due to issues with callback in combination with fork() on Centos.
See https://github.com/ffi/ffi/issues/621 .
It is caused by specialized handling of callback allocation on SE-Linux
protected systems through ffi_closure_alloc(), which is incompatible to fork().

Reverting this now, pending more investigation, how we can solve callback
allocation in a way that is compatible to both PaX and SE-Linux.