Move FFI::Platform::CPU from C to Ruby
The CPU detection was moved to C 10 years ago as a workaround to an
incorrect RbConfig::CONFIG['host_cpu'] value on MacOS.
This was in commit c53282af6 .
Although I can't reproduce the original environment, I guess that
this issue has been fixed years ago, so that the workaround isn't
necessary anymore.

Moving CPU detection to Ruby allows to use lib/ffi/platform.rb in
Rakefile before the C ext is built.