Fix library name mangling for non glibc Linux/UNIX
It incorrectly appended ".so" to lib names like "libglib-2.0.so.0" on musl based Alpine Linux.
A bit research showed, that the notation of the ABI number after ".so" is common on at least:
* glibc based Linux
* non-glibc Linux
* Freebsd
* Solaris
* Netbsd

So we change this to exclude only Windows and Macos from this rule.

Raised and authored by John Cupitt.

Fixes #727