Remove taint support
Ruby 2.7 deprecates taint and it no longer has an effect.
https://bugs.ruby-lang.org/issues/16131

The APIs related to taint cause deprecation warning message and the lack
of taint support should not cause a problem in previous Ruby versions.
So I removed that. But if need to taint support that to old versions
of Ruby, I will update PR.