Add SDK paths on macOS
MacOS removed `/usr/include` in recent versions of macOS.
Also pkg-config seems to be broken in conjunction with libffi.
That forced the usage of the internal libffi although macOS has a usable one itself.

This commit adds typical paths of the SDK which contain libffi.
Since it's possible to use non default SDK paths, there is another fallback to determine the path per "xcrun" command.

Fixes #757
Closes #765
Closes #758