Implement callback returning type :string
Callbacks returning a :string type were not supported so far.
It was possible to define such a callback, but the value returned was NULL in any case.
This implementation follows the semantics for passing :string parameters to functions as implemented in Call.c.
That means in particular usage of StringValueCStr(), so that no zero bytes are allowed in the string.

Special care has to be taken regarding lifetime of the returned text pointer.
The C pointer to the string is only valid until the next GC triggering function is called, unless the String object is hold in Ruby outside of the callback block.
Even when the object is stored outside of the Ruby block, the text pointer can move when GC.compact is called.
So :string returning callbacks can safely be used only, when the referenced text is processed immediately by the calling C function.

Resolves #751