Revert to msvcrt.dll as LIBC on MINGW
ffi-1.13.0 switched FFI::Library::LIBC from msvcrt.dll to ucrtbase.dll
as part of #779 in commit c67468366e63bcf84512837384049fbba7e4748c .

As described in #788 ucrtbase.dll has behavioural changes which
shouldn't be released as part of a minor version change of ffi.
While the change makes sense for mswin, we revert it for mingw.

Fixes #788