Fix possible UTF-8 error in ffi_lib rescue block
When a library can not be loaded is is tried to be interpret as a loader script.
This failed when the failing library is a binary file and has no valid UTF-8 content.
In this case the original LoadError was hidden.
Open as a binary file fixes this.

Also add a spec for loader script interpretation.

Fixes #792