Add support for new thread scheduler of Ruby-3.0
The feature is described in: https://bugs.ruby-lang.org/issues/16786

To avoid blocking the current ruby thread while calls to C, calls can be executed in a dedicated pthread.
This happens when the current thread has a scheduler assigned by Thread.current.scheduler= .

A pipe is used to signal the end of a call and the scheduler is invoked in order to wait for readability of the pipe.
This way the scheduler can yield to another fiber or do other work instead of blocking the thread until the C call finishs.

The current implementation does not yield any callbacks back to the calling thread.
Instead all callbacks invoked in this way are handled as asynchronous callbacks.
This means that each callback is executed in a dedicated ruby thread.