Check versions without accessing Gem
When running with --disable-gems, we still need to be able to
load FFI from stdlib. Accessing the Gem namespace here requires
that gems not be disabled.

This patch splits and compares the integer elements of version
directly rather than using Gem::Version to do that comparison.

See #763 for the original code.