Fix possible segfault in combination with fiddle or other libffi using gems
Fiddle can crash when used together with ffi and it's builtin libffi.
This happens because fiddle is linked to system libffi, but ffi is linked to builtin libffi.
Depending on which of these gems is loaded first, they link to the wrong runtime library.

This issue is similar to the issue in nokogiri: https://github.com/sparklemotion/nokogiri/issues/1959

Fixes #835