Name the async_cb_thread for easier debugging
I found this useful while reviewing datadog/dd-trace-rb#1371 -- we were
searching for specs that leaked threads, and saw a "mysterious" thread
with no stack that we couldn't exactly point out what created it or
why it existed.

By setting a name on the callback thread, it becomes immediately
obvious when looking at `Thread.list` what this thread is and where
it comes from.

(Thread naming has been around
[since Ruby 2.3](https://github.com/ruby/ruby/blob/v2_3_0/NEWS) which
means we can do it safely for every supported version)