Implement Write Barrier and dsize for FFI::Pointer
As well as FFI::MemoryPointer and FFI::AbstractMemory

Write barrier protected objects are allowed to be promoted to the old generation, which means they only get marked on major GC.

The downside is that the RB_BJ_WRITE macro MUST be used to set references, otherwise the referenced object may be garbaged collected.

This commit also implement a dsize function so that these instance report a more relevant size in various memory profilers.