Add support for using FFI in Ractor
All objects are shareable now when frozen.
All objects can be created in a non-main Ractor.

Typedefs are a global mutable state and are not accessable from Ractor other than the main Ractor.
So all Function, Struct, etc. must be defined in the main Ractor and can then be used in other Ractors.