Allow private release method for ManagedStruct and AutoPointer
FFI users might not want to expose `release` as a public class method, but want to keep it private.

Also re-enable GC tests that were disabled as part of #427, but make them opt-in per env var FFI_TEST_GC.

Derivations of Releaser are no longer necessary, so that they are removed now.