Avoid rb_class_new_instance() here, to avoid passing the method block
rb_class_new_instance() automatically passes the block forward to the initialize method.
That way the block of a callback is forwarded to the Struct initializer, like here in the specs:

      s = LibTest::S8F32S32.new
      ret = LibTest.testCallbackVrT { s }

Struct#initialize ignores the block, so that the code works, but it doesn't make sense.

This stood up when enabling rbs type checks, which complained about an unspecified block to Struct#initialize.