Avoid rb_class_new_instance() here, to avoid passing the method block
Similar to commit 5b4b4060727f7ade3842011350635370970ef650

rb_class_new_instance() automatically passes the block forward to the initialize method.
That way the block of an array function is forwarded to the Struct initializer.
Struct#initialize ignores the block, so that the code works, but it doesn't make sense.

This stood up when enabling rbs type checks, which complained about an unspecified block to Struct#initialize.