Remove libffi build dir after install
Previously if you installed the gem with `--disable-system-libffi`,
the gem would leave around the compiled libraries for ffi when
it no longer needed them.

This commits hooks into the Makefile install and removes
the build directory after the gem has successfully installed.

On my system, this appears to reduce the install usage for libffi 1.17.2
from 8.5 MB to 7.3 MB.